\name{ugir}
\alias{ugir}
\docType{package}
\title{Functions to support registry-based research at UGIR}
\description{
This is a collections of commonly used functions to work with the Swedish registers.
}
\details{
\tabular{ll}{
Package: \tab ugir\cr
Type: \tab Package\cr
Version: \tab 0.2.4\cr
Date: \tab 2015-02-24\cr
License: \tab For internal use only\cr
}
}
\author{Peter Konings

Maintainer: Peter Konings \email{peter.konings@ki.se}}
